third_party/WebKit/LayoutTests/inspector/tracing/timeline-time.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>

function simpleConsoleTime()
{
    console.time("a");
    console.timeEnd("a");
}

function nestedConsoleTime()
{
    console.time("a");
    {
        console.time("b");
        console.timeEnd("b");
        {
            console.time("c");
            {
                console.time("d");
                console.timeEnd("d");
            }
            console.timeEnd("c");
        }
    }
    console.timeEnd("a");
}


function unbalancedConsoleTime()
{
    console.time("a");
    console.time("b");
    console.timeEnd("a");
    console.timeEnd("b");
}

function consoleTimeWithoutConsoleTimeEnd()
{
    console.timeStamp("Foo");
    console.time("a");
    console.timeStamp("Bar");
    console.time("b");
    console.time("c");
    console.time("d");
    console.timeStamp("Baz");
    console.timeEnd("d");
}

function test()
{
    InspectorTest.runTestSuite([
        function testSimpleConsoleTime(next)
        {
            performActions("simpleConsoleTime()", next);
        },

        function testNestedConsoleTime(next)
        {
            performActions("nestedConsoleTime()", next);
        },

        function testUnbalancedConsoleTime(next)
        {
            performActions("unbalancedConsoleTime()", next);
        },

        function testConsoleTimeWithoutConsoleTimeEnd(next)
        {
            performActions("consoleTimeWithoutConsoleTimeEnd()", next);
        }
    ]);

    function performActions(actions, next)
    {
        var namesToDump = new Set(["FunctionCall", "ConsoleTime", "TimeStamp"]);
        function dumpName(event, level)
        {
            if (namesToDump.has(event.name))
                InspectorTest.addResult("----".repeat(level) + "> " + WebInspector.TimelineUIUtils.eventTitle(event));
        }
        function callback()
        {
            InspectorTest.walkTimelineEventTree(dumpName);
            next();
        }
        WebInspector.panels.timeline._captureJSProfileSetting.set(false);
        InspectorTest.evaluateWithTimeline(actions, InspectorTest.safeWrap(callback), true);
    }
}

</script>
</head>

<body onload="runTest()">
<p>Test nesting of time/timeEnd records on Timeline</p>

</body>
</html>
